Moving test server to standalone backend
