+++ backend/.gitignore
@@ -1,0 +1,8 @@
+# Specifies files to intentionally ignore when using Git
+# http://git-scm.com/docs/gitignore
+
+node_modules/
+*.swp
+.DS_Store
+Thumbs.db
+*.log
