+++ backend/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "bika_backend",
+  "description": "Bika: Aplicación híbrida para encontrar y reseñar bicicleterías.",
+  "author": "Camba Laboratorio de tecnología <info@camba.coop>",
+  "homepage": "https://github.com/Cambalab/bika",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Cambalab/bika.git"
+  },
+  "dependencies": {
+    "express": "^4.13.4",
+    "body-parser": "^1.14.2",
+    "compression": "^1.6.1",
+    "cors": "^2.7.1",
+    "lodash": "^4.6.1"
+  },
+  "devDependencies": {
+  }
+}
